+++ xSQLServerHelper.psm1
@@ -1,6 +1,9 @@
 # Load Localization Data
-Import-LocalizedData LocalizedData -filename xSQLServer.strings.psd1 -ErrorAction SilentlyContinue
-Import-LocalizedData USLocalizedData -filename xSQLServer.strings.psd1 -UICulture en-US -ErrorAction SilentlyContinue
+Import-Module -Name (Join-Path -Path (Join-Path -Path $PSScriptRoot `
+                                                -ChildPath 'DscResources') `
+                               -ChildPath 'CommonResourceHelper.psm1')
+
+$script:localizedData = Get-LocalizedData -ResourceName 'xSQLServerHelper' -ScriptRoot $PSScriptRoot
 
 <#
     .SYNOPSIS
@@ -27,7 +30,7 @@ function Connect-SQL
 
         [ValidateNotNull()]
         [System.String]
-        $SQLInstanceName = "MSSQLSERVER",
+        $SQLInstanceName = 'MSSQLSERVER',
 
         [ValidateNotNull()]
         [System.Management.Automation.PSCredential]
@@ -36,13 +39,13 @@ function Connect-SQL
 
     $null = [System.Reflection.Assembly]::LoadWithPartialName('Microsoft.SqlServer.Smo')
 
-    if ($SQLInstanceName -eq "MSSQLSERVER")
+    if ($SQLInstanceName -eq 'MSSQLSERVER')
     {
-        $connectSql = $SQLServer
+        $databaseEngineInstance = $SQLServer
     }
     else
     {
-        $connectSql = "$SQLServer\$SQLInstanceName"
+        $databaseEngineInstance = "$SQLServer\$SQLInstanceName"
     }
 
     if ($SetupCredential)
@@ -51,20 +54,21 @@ function Connect-SQL
         $sql.ConnectionContext.ConnectAsUser = $true
         $sql.ConnectionContext.ConnectAsUserPassword = $SetupCredential.GetNetworkCredential().Password
         $sql.ConnectionContext.ConnectAsUserName = $SetupCredential.GetNetworkCredential().UserName
-        $sql.ConnectionContext.ServerInstance = $connectSQL
+        $sql.ConnectionContext.ServerInstance = $databaseEngineInstance
         $sql.ConnectionContext.connect()
     }
     else
     {
-        $sql = New-Object Microsoft.SqlServer.Management.Smo.Server $connectSql
+        $sql = New-Object Microsoft.SqlServer.Management.Smo.Server $databaseEngineInstance
     }
 
-    if (!$sql)
+    if (-not $sql)
     {
-        throw "Failed connecting to SQL $connectSql"
+        $errorMessage = $script:localizedData.FailedToConnectToDatabaseEngineInstance -f $databaseEngineInstance
+        New-InvalidOperationException -Message $errorMessage
     }
 
-    New-VerboseMessage -Message "Connected to SQL $connectSql"
+    Write-Verbose -Message ($script:localizedData.ConnectedToDatabaseEngineInstance -f $databaseEngineInstance)
 
     return $sql
 }
@@ -137,14 +141,16 @@ function Connect-SQLAnalysis
         }
         else
         {
-            throw New-TerminatingError -ErrorType AnalysisServicesNoServerObject -ErrorCategory InvalidResult
+            $errorMessage = $script:localizedData.FailedToConnectToAnalysisServicesInstance -f $analysisServiceInstance
+            New-InvalidOperationException -Message $errorMessage
         }
 
-        Write-Verbose -Message "Connected to Analysis Services $analysisServiceInstance." -Verbose
+        Write-Verbose -Message ($script:localizedData.ConnectedToAnalysisServicesInstance -f $analysisServiceInstance)
     }
     catch
     {
-        throw New-TerminatingError -ErrorType AnalysisServicesFailedToConnect -FormatArgs @($analysisServiceInstance) -ErrorCategory ObjectNotFound -InnerException $_.Exception
+        $errorMessage = $script:localizedData.FailedToConnectToAnalysisServicesInstance -f $analysisServiceInstance
+        New-InvalidOperationException -Message $errorMessage -ErrorRecord $_
     }
 
     return $analysisServicesObject
@@ -189,22 +195,23 @@ function Register-SqlSmo
     )
 
     $sqlMajorVersion = Get-SqlInstanceMajorVersion -SQLInstanceName $SQLInstanceName
-    New-VerboseMessage -Message ('SQL major version is {0}.' -f $sqlMajorVersion)
 
-    if( -not $ApplicationDomain )
+    Write-Verbose -Message ($script:localizedData.SqlMajorVersion -f $sqlMajorVersion)
+
+    if ( -not $ApplicationDomain )
     {
         $applicationDomainName = $MyInvocation.MyCommand.ModuleName
-        New-VerboseMessage -Message ('Creating application domain ''{0}''.' -f $applicationDomainName)
+        Write-Verbose -Message ($script:localizedData.CreatingApplicationDomain -f $applicationDomainName)
         $applicationDomainObject = [System.AppDomain]::CreateDomain($applicationDomainName)
     }
     else
     {
-        New-VerboseMessage -Message ('Reusing application domain ''{0}''.' -f $ApplicationDomain.FriendlyName)
+        Write-Verbose -Message ($script:localizedData.ReusingApplicationDomain -f $ApplicationDomain.FriendlyName)
         $applicationDomainObject = $ApplicationDomain
     }
 
     $sqlSmoAssemblyName = "Microsoft.SqlServer.Smo, Version=$sqlMajorVersion.0.0.0, Culture=neutral, PublicKeyToken=89845dcd8080cc91"
-    New-VerboseMessage -Message ('Loading assembly ''{0}''.' -f $sqlSmoAssemblyName)
+    Write-Verbose -Message ($script:localizedData.LoadingAssembly -f $sqlSmoAssemblyName)
     $applicationDomainObject.Load($sqlSmoAssemblyName) | Out-Null
 
     return $applicationDomainObject
@@ -249,7 +256,7 @@ function Register-SqlWmiManagement
     )
 
     $sqlMajorVersion = Get-SqlInstanceMajorVersion -SQLInstanceName $SQLInstanceName
-    New-VerboseMessage -Message ('SQL major version is {0}.' -f $sqlMajorVersion)
+    Write-Verbose -Message ($script:localizedData.SqlMajorVersion -f $sqlMajorVersion)
 
     <#
         Must register Microsoft.SqlServer.Smo first because that is a
@@ -266,7 +273,7 @@ function Register-SqlWmiManagement
     }
 
     $sqlSqlWmiManagementAssemblyName = "Microsoft.SqlServer.SqlWmiManagement, Version=$sqlMajorVersion.0.0.0, Culture=neutral, PublicKeyToken=89845dcd8080cc91"
-    New-VerboseMessage -Message ('Loading assembly ''{0}''.' -f $sqlSqlWmiManagementAssemblyName)
+    Write-Verbose -Message ($script:localizedData.LoadingAssembly -f $sqlSqlWmiManagementAssemblyName)
     $applicationDomainObject.Load($sqlSqlWmiManagementAssemblyName) | Out-Null
 
     return $applicationDomainObject
@@ -290,7 +297,7 @@ function Unregister-SqlAssemblies
         $ApplicationDomain
     )
 
-    New-VerboseMessage -Message ('Unloading application domain ''{0}''.' -f $ApplicationDomain.FriendlyName)
+    Write-Verbose -Message ($script:localizedData.UnloadingApplicationDomain -f $ApplicationDomain.FriendlyName)
     [System.AppDomain]::Unload($ApplicationDomain)
 }
 
@@ -320,7 +327,8 @@ function Get-SqlInstanceMajorVersion
 
     if (-not $sqlVersion)
     {
-        throw New-TerminatingError -ErrorType SqlServerVersionIsInvalid -FormatArgs @($SQLInstanceName) -ErrorCategory InvalidResult
+        $errorMessage = $script:localizedData.SqlServerVersionIsInvalid -f $SQLInstanceName
+        New-InvalidResultException -Message $errorMessage
     }
 
     [System.UInt16] $sqlMajorVersionNumber = $sqlVersion.Split('.')[0]
@@ -331,6 +339,13 @@ function Get-SqlInstanceMajorVersion
 <#
     .SYNOPSIS
         Returns a localized error message.
+
+        This helper function is obsolete, should use new helper functions.
+        https://github.com/PowerShell/xSQLServer/blob/dev/CONTRIBUTING.md#localization
+        https://github.com/PowerShell/xSQLServer/blob/dev/DSCResources/CommonResourceHelper.psm1
+
+        Strings in this function has not been localized since this helper function should be removed
+        when all resources has moved over to the new localization,
 
     .PARAMETER ErrorType
         String containing the key of the localized error message.
@@ -376,11 +391,11 @@ function New-TerminatingError
         $InnerException = $null
     )
 
-    $errorMessage = $LocalizedData.$ErrorType
+    $errorMessage = $script:localizedData.$ErrorType
 
     if(!$errorMessage)
     {
-        $errorMessage = ($LocalizedData.NoKeyFound -f $ErrorType)
+        $errorMessage = ($script:localizedData.NoKeyFound -f $ErrorType)
 
         if(!$errorMessage)
         {
@@ -411,7 +426,7 @@ function New-TerminatingError
         $errorId = $ErrorType
     }
 
-    Write-Verbose -Message "$($USLocalizedData.$ErrorType -f $FormatArgs) | ErrorType: $errorId"
+    Write-Verbose -Message "$($script:localizedData.$ErrorType -f $FormatArgs) | ErrorType: $errorId"
 
     $exception = New-Object System.Exception $errorMessage, $InnerException
     $errorRecord = New-Object System.Management.Automation.ErrorRecord $exception, $errorId, $ErrorCategory, $TargetObject
@@ -422,6 +437,13 @@ function New-TerminatingError
 <#
     .SYNOPSIS
         Displays a localized warning message.
+
+        This helper function is obsolete, should use Write-Warning together with individual resource
+        localization strings.
+        https://github.com/PowerShell/xSQLServer/blob/dev/CONTRIBUTING.md#localization
+
+        Strings in this function has not been localized since this helper function should be removed
+        when all resources has moved over to the new localization,
 
     .PARAMETER WarningType
         String containing the key of the localized warning message.
@@ -444,7 +466,7 @@ function New-WarningMessage
     )
 
     ## Attempt to get the string from the localized data
-    $warningMessage = $LocalizedData.$WarningType
+    $warningMessage = $script:localizedData.$WarningType
 
     ## Ensure there is a message present in the localization file
     if (!$warningMessage)
@@ -470,6 +492,13 @@ function New-WarningMessage
 <#
     .SYNOPSIS
     Displays a standardized verbose message.
+
+    This helper function is obsolete, should use Write-Verbose together with individual resource
+    localization strings.
+    https://github.com/PowerShell/xSQLServer/blob/dev/CONTRIBUTING.md#localization
+
+    Strings in this function has not been localized since this helper function should be removed
+    when all resources has moved over to the new localization,
 
     .PARAMETER Message
     String containing the key of the localized warning message.
@@ -525,13 +554,14 @@ function Test-SQLDscParameterState
         -and ($DesiredValues.GetType().Name -ne "CimInstance") `
         -and ($DesiredValues.GetType().Name -ne "PSBoundParametersDictionary"))
     {
-        throw "Property 'DesiredValues' in Test-SQLDscParameterState must be either a " + `
-              "Hash table, CimInstance or PSBoundParametersDictionary. Type detected was $($DesiredValues.GetType().Name)"
+        $errorMessage = $script:localizedData.PropertyTypeInvalidForDesiredValues -f $($DesiredValues.GetType().Name)
+        New-InvalidArgumentException -ArgumentName 'DesiredValues' -Message $errorMessage
     }
 
     if (($DesiredValues.GetType().Name -eq "CimInstance") -and ($null -eq $ValuesToCheck))
     {
-        throw "If 'DesiredValues' is a CimInstance then property 'ValuesToCheck' must contain a value"
+        $errorMessage = $script:localizedData.PropertyTypeInvalidForValuesToCheck
+        New-InvalidArgumentException -ArgumentName 'ValuesToCheck' -Message $errorMessage
     }
 
     if (($null -eq $ValuesToCheck) -or ($ValuesToCheck.Count -lt 1))
@@ -544,14 +574,14 @@ function Test-SQLDscParameterState
     }
 
     $keyList | ForEach-Object -Process {
-        if (($_ -ne "Verbose"))
+        if (($_ -ne 'Verbose'))
         {
             if (($CurrentValues.ContainsKey($_) -eq $false) `
             -or ($CurrentValues.$_ -ne $DesiredValues.$_) `
             -or (($DesiredValues.GetType().Name -ne 'CimInstance' -and $DesiredValues.ContainsKey($_) -eq $true) -and ($null -ne $DesiredValues.$_ -and $DesiredValues.$_.GetType().IsArray)))
             {
-                if ($DesiredValues.GetType().Name -eq "HashTable" -or `
-                    $DesiredValues.GetType().Name -eq "PSBoundParametersDictionary")
+                if ($DesiredValues.GetType().Name -eq 'HashTable' -or `
+                    $DesiredValues.GetType().Name -eq 'PSBoundParametersDictionary')
                 {
                     $checkDesiredValue = $DesiredValues.ContainsKey($_)
                 }
@@ -577,11 +607,7 @@ function Test-SQLDscParameterState
                         if (($CurrentValues.ContainsKey($fieldName) -eq $false) `
                         -or ($null -eq $CurrentValues.$fieldName))
                         {
-                            New-VerboseMessage -Message ("Expected to find an array value for " + `
-                                                         "property $fieldName in the current " + `
-                                                         "values, but it was either not present or " + `
-                                                         "was null. This has caused the test method " + `
-                                                         "to return false.")
+                            Write-Verbose -Message ($script:localizedData.PropertyValidationError -f $fieldName)
 
                             $returnValue = $false
                         }
@@ -591,12 +617,10 @@ function Test-SQLDscParameterState
                                                            -DifferenceObject $DesiredValues.$fieldName
                             if ($null -ne $arrayCompare)
                             {
-                                New-VerboseMessage -Message ("Found an array for property $fieldName " + `
-                                                             "in the current values, but this array " + `
-                                                             "does not match the desired state. " + `
-                                                             "Details of the changes are below.")
+                                Write-Verbose -Message ($script:localizedData.PropertiesDoesNotMatch -f $fieldName)
+
                                 $arrayCompare | ForEach-Object -Process {
-                                    New-VerboseMessage -Message "$($_.InputObject) - $($_.SideIndicator)"
+                                    Write-Verbose -Message ($script:localizedData.PropertyThatDoesNotMatch -f $_.InputObject, $_.SideIndicator)
                                 }
 
                                 $returnValue = $false
@@ -607,43 +631,42 @@ function Test-SQLDscParameterState
                     {
                         switch ($desiredType.Name)
                         {
-                            "String" {
+                            'String' {
                                 if (-not [System.String]::IsNullOrEmpty($CurrentValues.$fieldName) -or `
                                     -not [System.String]::IsNullOrEmpty($DesiredValues.$fieldName))
                                 {
-                                    New-VerboseMessage -Message ("String value for property $fieldName does not match. " + `
-                                                                 "Current state is '$($CurrentValues.$fieldName)' " + `
-                                                                 "and desired state is '$($DesiredValues.$fieldName)'")
+                                    Write-Verbose -Message ($script:localizedData.ValueOfTypeDoesNotMatch `
+                                        -f $desiredType.Name, $fieldName, $($CurrentValues.$fieldName), $($DesiredValues.$fieldName))
 
                                     $returnValue = $false
                                 }
                             }
-                            "Int32" {
+
+                            'Int32' {
                                 if (-not ($DesiredValues.$fieldName -eq 0) -or `
                                     -not ($null -eq $CurrentValues.$fieldName))
                                 {
-                                    New-VerboseMessage -Message ("Int32 value for property " + "$fieldName does not match. " + `
-                                                                 "Current state is " + "'$($CurrentValues.$fieldName)' " + `
-                                                                 "and desired state is " + "'$($DesiredValues.$fieldName)'")
+                                    Write-Verbose -Message ($script:localizedData.ValueOfTypeDoesNotMatch `
+                                        -f $desiredType.Name, $fieldName, $($CurrentValues.$fieldName), $($DesiredValues.$fieldName))
 
                                     $returnValue = $false
                                 }
                             }
-                            "Int16" {
+
+                            'Int16' {
                                 if (-not ($DesiredValues.$fieldName -eq 0) -or `
                                     -not ($null -eq $CurrentValues.$fieldName))
                                 {
-                                    New-VerboseMessage -Message ("Int32 value for property " + "$fieldName does not match. " + `
-                                                                 "Current state is " + "'$($CurrentValues.$fieldName)' " + `
-                                                                 "and desired state is " + "'$($DesiredValues.$fieldName)'")
+                                    Write-Verbose -Message ($script:localizedData.ValueOfTypeDoesNotMatch `
+                                        -f $desiredType.Name, $fieldName, $($CurrentValues.$fieldName), $($DesiredValues.$fieldName))
 
                                     $returnValue = $false
                                 }
                             }
+
                             default {
-                                Write-Warning -Message ("Unable to compare property $fieldName " + `
-                                                             "as the type ($($desiredType.Name)) is " + `
-                                                             "not handled by the Test-SQLDscParameterState cmdlet")
+                                Write-Warning -Message ($script:localizedData.ValueOfTypeDoesNotMatch `
+                                    -f $fieldName, $desiredType.Name)
 
                                 $returnValue = $false
                             }
@@ -669,11 +692,11 @@ function Import-SQLPSModule
     $module = (Get-Module -FullyQualifiedName 'SqlServer' -ListAvailable).Name
     if ($module)
     {
-        New-VerboseMessage -Message 'Preferred module SqlServer found.'
+        Write-Verbose -Message ($script:localizedData.PreferredModuleFound)
     }
     else
     {
-        New-VerboseMessage -Message 'Module SqlServer not found, trying to use older SQLPS module.'
+        Write-Verbose -Message ($script:localizedData.PreferredModuleNotFound)
         $module = (Get-Module -FullyQualifiedName 'SQLPS' -ListAvailable).Name
     }
 
@@ -681,10 +704,10 @@ function Import-SQLPSModule
     {
         try
         {
-            Write-Debug -Message 'SQLPS module changes CWD to SQLSERVER:\ when loading, pushing location to pop it when module is loaded.'
+            Write-Debug -Message ($script:localizedData.DebugMessagePushingLocation)
             Push-Location
 
-            New-VerboseMessage -Message ('Importing {0} module.' -f $module)
+            Write-Verbose -Message ($script:localizedData.ImportingPowerShellModule -f $module)
 
             <#
                 SQLPS has unapproved verbs, disable checking to ignore Warnings.
@@ -692,21 +715,23 @@ function Import-SQLPSModule
             #>
             Import-Module -Name $module -DisableNameChecking -Verbose:$False -ErrorAction Stop
 
-            Write-Debug -Message ('Module {0} imported.' -f $module)
+            Write-Debug -Message ($script:localizedData.DebugMessageImportedPowerShellModule -f $module)
         }
         catch
         {
-            throw New-TerminatingError -ErrorType FailedToImportSqlModule -FormatArgs @($module) -ErrorCategory InvalidOperation -InnerException $_.Exception
+            $errorMessage = $script:localizedData.FailedToImportPowerShellSqlModule -f $module
+            New-InvalidOperationException -Message $errorMessage -ErrorRecord $_
         }
         finally
         {
-            Write-Debug -Message 'Popping location back to what it was before importing SQLPS module.'
+            Write-Debug -Message ($script:localizedData.DebugMessagePoppingLocation)
             Pop-Location
         }
     }
     else
     {
-        throw New-TerminatingError -ErrorType SqlModuleNotFound -ErrorCategory InvalidOperation -InnerException $_.Exception
+        $errorMessage = $script:localizedData.PowerShellSqlModuleNotFound
+        New-InvalidOperationException -Message $errorMessage
     }
 }
 
@@ -755,49 +780,51 @@ function Restart-SqlService
 
     if ($serverObject.IsClustered)
     {
-        ## Get the cluster resources
-        New-VerboseMessage -Message 'Getting cluster resource for SQL Server'
+        # Get the cluster resources
+        Write-Verbose -Message ($script:localizedData.GetSqlServerClusterResources)
         $sqlService = Get-CimInstance -Namespace root/MSCluster -ClassName MSCluster_Resource -Filter "Type = 'SQL Server'" |
-                        Where-Object { $_.PrivateProperties.InstanceName -eq $serverObject.ServiceName }
+                        Where-Object -FilterScript { $_.PrivateProperties.InstanceName -eq $serverObject.ServiceName }
 
-        New-VerboseMessage -Message 'Getting active cluster resource SQL Server Agent'
+        Write-Verbose -Message ($script:localizedData.GetSqlAgentClusterResource)
         $agentService = $sqlService | Get-CimAssociatedInstance -ResultClassName MSCluster_Resource |
-                            Where-Object { ($_.Type -eq "SQL Server Agent") -and ($_.State -eq 2) }
+                            Where-Object -FilterScript { ($_.Type -eq 'SQL Server Agent') -and ($_.State -eq 2) }
 
-        ## Build a listing of resources being acted upon
+        # Build a listing of resources being acted upon
         $resourceNames = @($sqlService.Name, ($agentService | Select-Object -ExpandProperty Name)) -join ","
 
-        ## Stop the SQL Server and dependent resources
-        New-VerboseMessage -Message "Bringing the SQL Server resources $resourceNames offline."
+        # Stop the SQL Server and dependent resources
+        Write-Verbose -Message ($script:localizedData.BringClusterResourcesOffline -f $resourceNames)
         $sqlService | Invoke-CimMethod -MethodName TakeOffline -Arguments @{ Timeout = $Timeout }
 
-        ## Start the SQL server resource
-        New-VerboseMessage -Message 'Bringing the SQL Server resource back online.'
+        # Start the SQL server resource
+        Write-Verbose -Message ($script:localizedData.BringSqlServerClusterResourcesOnline)
         $sqlService | Invoke-CimMethod -MethodName BringOnline -Arguments @{ Timeout = $Timeout }
 
-        ## Start the SQL Agent resource
+        # Start the SQL Agent resource
         if ($agentService)
         {
-            New-VerboseMessage -Message 'Bringing the SQL Server Agent resource online.'
+            Write-Verbose -Message ($script:localizedData.BringSqlServerAgentClusterResourcesOnline)
             $agentService | Invoke-CimMethod -MethodName BringOnline -Arguments @{ Timeout = $Timeout }
         }
     }
     else
     {
-        New-VerboseMessage -Message 'Getting SQL Service information'
+        Write-Verbose -Message ($script:localizedData.GetSqlServerService)
         $sqlService = Get-Service -DisplayName "SQL Server ($($serverObject.ServiceName))"
 
-        ## Get all dependent services that are running.
-        ## There are scenarios where an automatic service is stopped and should not be restarted automatically.
-        $agentService = $sqlService.DependentServices | Where-Object { $_.Status -eq "Running" }
+        <#
+            Get all dependent services that are running.
+            There are scenarios where an automatic service is stopped and should not be restarted automatically.
+        #>
+        $agentService = $sqlService.DependentServices | Where-Object -FilterScript { $_.Status -eq 'Running' }
 
-        ## Restart the SQL Server service
-        New-VerboseMessage -Message 'SQL Server service restarting'
+        # Restart the SQL Server service
+        Write-Verbose -Message ($script:localizedData.RestartSqlServerService)
         $sqlService | Restart-Service -Force
 
-        ## Start dependent services
+        # Start dependent services
         $agentService | ForEach-Object {
-            New-VerboseMessage -Message "Starting $($_.DisplayName)"
+            Write-Verbose -Message ($script:localizedData.StartingDependentService -f $_.DisplayName)
             $_ | Start-Service
         }
     }
@@ -865,7 +892,8 @@ function Invoke-Query
         }
         catch
         {
-            throw New-TerminatingError -ErrorType ExecuteQueryWithResultsFailed -FormatArgs $Database -ErrorCategory NotSpecified
+            $errorMessage = $script:localizedData.ExecuteQueryWithResultsFailed -f $Database
+            New-InvalidOperationException -Message $errorMessage -ErrorRecord $_
         }
     }
     else
@@ -876,7 +904,8 @@ function Invoke-Query
         }
         catch
         {
-            throw New-TerminatingError -ErrorType ExecuteNonQueryFailed -FormatArgs $Database -ErrorCategory NotSpecified
+            $errorMessage = $script:localizedData.ExecuteNonQueryFailed -f $Database
+            New-InvalidOperationException -Message $errorMessage -ErrorRecord $_
         }
     }
 
@@ -907,7 +936,8 @@ function Update-AvailabilityGroupReplica
     }
     catch
     {
-        throw New-TerminatingError -ErrorType AlterAvailabilityGroupReplicaFailed -FormatArgs $AvailabilityGroupReplica.Name -ErrorCategory OperationStopped
+        $errorMessage = $script:localizedData.AlterAvailabilityGroupReplicaFailed -f $AvailabilityGroupReplica.Name
+        New-InvalidOperationException -Message $errorMessage -ErrorRecord $_
     }
     finally
     {
@@ -941,7 +971,7 @@ function Test-LoginEffectivePermissions
     # Assume the permissions are not present
     $permissionsPresent = $false
 
-    $invokeQueryParams = @{
+    $invokeQueryParameters = @{
         SQLServer = $SQLServer
         SQLInstanceName = $SQLInstanceName
         Database = 'master'
@@ -955,9 +985,9 @@ function Test-LoginEffectivePermissions
         REVERT
     "
 
-    New-VerboseMessage -Message "Getting the effective permissions for the login '$LoginName' on '$sqlInstanceName'."
+    Write-Verbose -Message ($script:localizedData.GetEffectivePermissionForLogin -f $LoginName, $sqlInstanceName)
 
-    $loginEffectivePermissionsResult = Invoke-Query @invokeQueryParams -Query $queryToGetEffectivePermissionsForLogin
+    $loginEffectivePermissionsResult = Invoke-Query @invokeQueryParameters -Query $queryToGetEffectivePermissionsForLogin
     $loginEffectivePermissions = $loginEffectivePermissionsResult.Tables.Rows.permission_name
 
     if ( $null -ne $loginEffectivePermissions )
@@ -981,7 +1011,7 @@ function Test-LoginEffectivePermissions
 
     .PARAMETER SQLServer
         The hostname of the server that hosts the SQL instance.
-    
+
     .PARAMETER SQLInstanceName
         The name of the SQL instance that hosts the availability group.
 
@@ -1019,7 +1049,7 @@ function Test-AvailabilityReplicaSeedingModeAutomatic
     $availabilityReplicaSeedingModeAutomatic = $false
 
     $serverObject = Connect-SQL -SQLServer $SQLServer -SQLInstanceName $SQLInstanceName
-    
+
     # Only check the seeding mode if this is SQL 2016 or newer
     if ( $serverObject.Version -ge 13 )
     {
@@ -1029,7 +1059,7 @@ function Test-AvailabilityReplicaSeedingModeAutomatic
             Database = 'master'
             WithResults = $true
         }
-        
+
         $queryToGetSeedingMode = "
             SELECT seeding_mode_desc
             FROM sys.availability_replicas ar
@@ -1055,7 +1085,7 @@ function Test-AvailabilityReplicaSeedingModeAutomatic
 
     .PARAMETER ServerObject
         The current server object connection.
-    
+
     .PARAMETER AvailabilityGroup
         The availability group object used to find the primary replica server name.
 #>
@@ -1066,14 +1096,14 @@ function Get-PrimaryReplicaServerObject
         [Parameter(Mandatory = $true)]
         [Microsoft.SqlServer.Management.Smo.Server]
         $ServerObject,
-        
+
         [Parameter(Mandatory = $true)]
         [Microsoft.SqlServer.Management.Smo.AvailabilityGroup]
         $AvailabilityGroup
     )
 
     $primaryReplicaServerObject = $serverObject
-    
+
     # Determine if we're connected to the primary replica
     if ( ( $AvailabilityGroup.PrimaryReplicaServerName -ne $serverObject.DomainInstanceName ) -and ( -not [System.String]::IsNullOrEmpty($AvailabilityGroup.PrimaryReplicaServerName) ) )
     {
@@ -1105,7 +1135,7 @@ function Test-ImpersonatePermissions
         LoginName = $ServerObject.ConnectionContext.TrueLogin
         Permissions = @('IMPERSONATE ANY LOGIN')
     }
-    
+
     $impersonatePermissionsPresent = Test-LoginEffectivePermissions @testLoginEffectivePermissionsParams
 
     if ( -not $impersonatePermissionsPresent )
@@ -1122,7 +1152,7 @@ function Test-ImpersonatePermissions
 
     .PARAMETER FullSQLInstanceName
         The full SQL instance name string to be split.
-    
+
     .OUTPUTS
         Hashtable with the properties SQLServer and SQLInstanceName.
 #>