+++ CHANGELOG.md
@@ -51,6 +51,7 @@
   - Test-SQLDscParameterState helper function can now correctly pass a CimInstance as DesiredValue.
   - Test-SQLDscParameterState helper function will now output a warning message
     if the value type of a desired value is not supported.
+  - Added localization to helper functions (issue #641)
 - Changes to xSQLServerSetup
   - BREAKING CHANGE: Replaced StartWin32Process helper function with the cmdlet Start-Process (issue #41, #93 and #126).
   - BREAKING CHANGE: The parameter SetupCredential has been removed since it is no longer needed. This is because the resource now support the built-in PsDscRunAsCredential.